predictor/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
black @ git+git://github.com/psf/black@b3ceb293d9e69295a190fed93517cbe1b7372154
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.1
soupsieve==2.2
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.3
Werkzeug==1.0.1
